Task/14 user interface 1

See merge request ussass/timetable-viewer-system!7

+++ src/main/java/ru/trofimov/timetableviewersystem/dao/mapper/StudentMapper.java
@@ -15,7 +15,7 @@ public class StudentMapper implements RowMapper<Student> {
                 resultSet.getString("last_name")
         );
         student.setId(resultSet.getLong("student_id"));
-        student.setGroupId(resultSet.getInt("group_id"));
+        student.setGroupId(resultSet.getLong("group_id"));
 
         return student;
     }
